renamed Board to Grid
